fixed types issue

+++ util/meter.py
@@ -41,12 +41,12 @@ class StatsMeter(Meter):
         Keyword Arguments:
             iterable {[iterable} -- Values to initialize (default: {None})
         """
-        self.n: int = 0
+        self.n: float = 0
         self.mean: Numeric = 0.0
         self.S: Numeric = 0.0
         super().__init__(iterable)
 
-    def add(self, datum: Numeric, n: int = 1):
+    def add(self, datum: Numeric, n: float = 1):
         """Add a single datum
 
         Internals are updated using Welford's method
@@ -108,7 +108,7 @@ class StatsMeter(Meter):
         return np.sqrt(self.variance)
 
     @staticmethod
-    def from_values(n: int, mean: float, std: float) -> "StatsMeter":
+    def from_values(n: float, mean: float, std: float) -> "StatsMeter":
         stats = StatsMeter()
         stats.n = n
         stats.mean = mean
@@ -116,7 +116,7 @@ class StatsMeter(Meter):
         return stats
 
     @staticmethod
-    def from_raw_values(n: int, mean: float, S: float) -> "StatsMeter":
+    def from_raw_values(n: float, mean: float, S: float) -> "StatsMeter":
         stats = StatsMeter()
         stats.n = n
         stats.mean = mean
